Move setup tools to a module.
