+++ ArchSetup
@@ -17,7 +17,7 @@
 # along with ArchSetup.  If not, see <http://www.gnu.org/licenses/>.
 
 import logging
-from SetupTools import SetupTools
+from SetupTools.Setuptools import SetupTools
 from Interface.Interface import Interface
 from Interface.WelcomeWindow import WelcomeWindow
 from Interface.FinishWindow import FinishWindow
