+++ SetupTools/Keyboard.py
@@ -16,14 +16,9 @@
 # You should have received a copy of the GNU General Public License
 # along with ArchSetup.  If not, see <http://www.gnu.org/licenses/>.
 
-import os
-import logging
-import subprocess
-
-class SetupTools:
+class Keyboard:
     def __init__(self):
-        if os.geteuid() != 0:
-            logging.warning('No root permissions...')
+        pass
 
     def list_keyboard_layouts(self):
         kbdlist=subprocess.check_output(["localectl", "list-keymaps"]).decode().split('\n')
@@ -36,27 +31,4 @@ class SetupTools:
 
     def set_keyboard_layout():
         print("[D] SetupTools.set_keyboard_layout(): Not implemented!")
-
-    def list_console_fonts(self):
-        fntlist=subprocess.check_output("(cd /usr/share/kbd/consolefonts/ && find . -type f -name '*.psf.gz' -o -name '*.psfu.gz') | sed 's/.\\/\\(.*\\).psf.gz/\\1/' | sed 's/.\\/\\(.*\\).psfu.gz/\\1/'", shell=True).decode().split('\n')
-        fntlist.remove('')
-        return fntlist
-
-    def load_console_font():
-        #setfont $font
-        print("[D] SetupTools.load_console_font(): Not implemented!")
-
-    def set_console_font():
-        print("[D] SetupTools.set_console_font(): Not implemented!")
-
-    def generate_locale():
-        #locale-gen
-        print("[D] SetupTools.generate_locale(): Not implemented!")
-
-    def check_internet():
-        #ping -c 5 www.google.com > /dev/null
-        print("[D] SetupTools.check_internet(): Not implemented!")
-
-    def add_user():
-        print("[D] SetupTools.add_user(): Not implemented!")
